+++ CHANGELOG.md
@@ -24,6 +24,8 @@ For change items without a ticket, the following format is used:
 ### Major Changes
 1. **2021-06-25**:
    1. \#8 Remove deprecated constructor ProjectRenderer.
+   1. \#8 Provide path `/net/splitcells/website/layout/build` in server,
+      in order to refresh project layout.
 ### Minor Changes
 1. **2021-06-25**:
    1. \#8 Create index file, for standard analysis.